refactor: group similar operators into 1 class (#316)
Groups operators together than are able to use 1 generic function with
different arguments passed in.


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1211219212741037
