+++ cdisc_rules_engine/check_operators/sql/base_sql_operator.py
@@ -275,106 +275,6 @@ class BaseSqlOperator:
         return series_to_validate.eq(comparison_data).astype(bool)"""
         raise NotImplementedError("_check_equality_of_string_part check_operator not implemented")
 
-    def _date_comparison(self, other_value: dict, operator: str):
-        """
-        Performs date comparison operations in PostgreSQL.
-        Handles date component extraction and comparison.
-        """
-        target_column = self.replace_prefix(other_value.get("target")).lower()
-        comparator = other_value.get("comparator")
-        value_is_literal = other_value.get("value_is_literal", False)
-        date_component = other_value.get("date_component")
-
-        if isinstance(comparator, str) and not value_is_literal:
-            comparator = self.replace_prefix(comparator).lower()
-
-        component_suffix = f"_{date_component}" if date_component else ""
-        cache_key = f"{target_column}{operator}{comparator}{component_suffix}"
-
-        def sql():
-            if date_component:
-                component_map = {
-                    "year": "YEAR",
-                    "month": "MONTH",
-                    "day": "DAY",
-                    "hour": "HOUR",
-                    "minute": "MINUTE",
-                    "second": "SECOND",
-                    "microsecond": "MICROSECONDS",
-                }
-                pg_component = component_map.get(date_component, "EPOCH")
-
-                if value_is_literal:
-                    return f"""CASE WHEN
-                        {self.replace_prefix(target_column)} IS NOT NULL
-                        AND {self.replace_prefix(target_column)} != ''
-                        AND EXTRACT({pg_component} FROM CAST({self.replace_prefix(target_column)} AS TIMESTAMP))
-                            {operator}
-                        EXTRACT({pg_component} FROM CAST('{comparator}' AS TIMESTAMP))
-                        THEN true
-                        ELSE false
-                        END"""
-                else:
-                    return f"""CASE WHEN
-                        {self.replace_prefix(target_column)} IS NOT NULL
-                        AND {self.replace_prefix(target_column)} != ''
-                        AND {self.replace_prefix(comparator)} IS NOT NULL
-                        AND {self.replace_prefix(comparator)} != ''
-                        AND EXTRACT({pg_component} FROM CAST({self.replace_prefix(target_column)} AS TIMESTAMP))
-                            {operator}
-                        EXTRACT({pg_component} FROM CAST({self.replace_prefix(comparator)} AS TIMESTAMP))
-                        THEN true
-                        ELSE false
-                        END"""
-            else:
-                if value_is_literal:
-                    return f"""CASE WHEN
-                        {target_column} IS NOT NULL
-                        AND {target_column} != ''
-                        AND CAST({target_column} AS TIMESTAMP)
-                            {operator}
-                        CAST('{comparator}' AS TIMESTAMP)
-                        THEN true
-                        ELSE false
-                        END"""
-                else:
-                    return f"""CASE WHEN
-                        {target_column} IS NOT NULL
-                        AND {target_column} != ''
-                        AND {comparator} IS NOT NULL
-                        AND {comparator} != ''
-                        AND CAST({target_column} AS TIMESTAMP)
-                            {operator}
-                        CAST({comparator} AS TIMESTAMP)
-                        THEN true
-                        ELSE false
-                        END"""
-
-        return self._do_check_operator(cache_key, sql)
-
-    def _numeric_comparison(
-        self,
-        other_value: dict,
-        operator: str,
-    ):
-        target_column = self.replace_prefix(other_value.get("target"))
-        comparator = (
-            other_value.get("comparator").lower()
-            if isinstance(other_value.get("comparator"), str)
-            else other_value.get("comparator")
-        )
-
-        def sql():
-            return f"""CASE WHEN
-                            CAST({target_column} AS NUMERIC)
-                                {operator}
-                            CAST({comparator} AS NUMERIC) THEN true
-                        ELSE false
-                        END
-                        """
-
-        return self._do_check_operator(f"{target_column}{operator}{comparator}", sql)
-
     def _series_is_in(self, target, comparison_data):
         return np.where(comparison_data.isin(target), True, False)
