--- cdisc_rules_engine/check_operators/sql/date_not_equal_to_operator.py
@@ -1,9 +1,0 @@
-from .base_sql_operator import BaseSqlOperator
-
-
-class DateNotEqualToOperator(BaseSqlOperator):
-    """Operator for date inequality comparisons."""
-
-    def execute_operator(self, other_value):
-        """Check if target date does not equal comparator date"""
-        return self._date_comparison(other_value, "!=")
