--- cdisc_rules_engine/check_operators/sql/greater_than_operator.py
@@ -1,8 +1,0 @@
-from .base_sql_operator import BaseSqlOperator
-
-
-class GreaterThanOperator(BaseSqlOperator):
-    """Operator for numeric greater-than comparisons."""
-
-    def execute_operator(self, other_value):
-        return self._numeric_comparison(other_value, ">")
