--- cdisc_rules_engine/check_operators/sql/less_than_or_equal_to_operator.py
@@ -1,8 +1,0 @@
-from .base_sql_operator import BaseSqlOperator
-
-
-class LessThanOrEqualToOperator(BaseSqlOperator):
-    """Operator for numeric less-than-or-equal-to comparisons."""
-
-    def execute_operator(self, other_value):
-        return self._numeric_comparison(other_value, "<=")
